Separate matrix and vector tests and add comments to SVD tests
